requirements.txt
jieba
x_transformers
librosa==0.10.2
pytorch-lightning>=2.4
cn2an
g2p_en
pypinyin
jieba_fast
pyopenjtalk
opencc
psutil
torchcodec
split-lang
wordsegment
fastapi[standard]>=0.115.2
fast_langdetect>=0.3.1
transformers
onnxruntime-gpu
matplotlib
tensorboard